Fix spec executions in group to run as in separate
When running gui_spec.rb in separate it works.
When running it with the other tests it fails.
The problem was that gui.rb was required before the stubing of Gosu::Window.
So, now we save the already created Gosu and Gui objects to restore them at the end.
We stub Gosu::Window and reload gui.rb so that the new Gui descend from our FakeGosuWindow.
Them we save this new Gui on FakeGui to use them on the tests without have to deal with display issues.
At the end we restore Gui and Gosu to their initial state.